Update json generation from market stack api

+++ FuzzyTrader.Contracts/External/MartketStackEndOfDayResponse.cs
@@ -12,17 +12,17 @@ namespace FuzzyTrader.Contracts.External
 
     public class MarketStackData
     {
-        public double? open { get; set; }
-        public double? high { get; set; }
-        public double? low { get; set; }
-        public double? close { get; set; }
-        public double? volume { get; set; }
-        public double? adj_high { get; set; }
-        public double? adj_low { get; set; }
-        public double? adj_close { get; set; }
-        public double? adj_open { get; set; }
-        public double? adj_volume { get; set; }
-        public double? split_factor { get; set; }
+        public decimal? open { get; set; }
+        public decimal? high { get; set; }
+        public decimal? low { get; set; }
+        public decimal? close { get; set; }
+        public decimal? volume { get; set; }
+        public decimal? adj_high { get; set; }
+        public decimal? adj_low { get; set; }
+        public decimal? adj_close { get; set; }
+        public decimal? adj_open { get; set; }
+        public decimal? adj_volume { get; set; }
+        public decimal? split_factor { get; set; }
         public string symbol { get; set; }
         public string exchange { get; set; }
         public string date { get; set; }
